# docs/api.yaml
openapi: 3.0.1
info:
  title: 银行管理系统
  description: 银行管理系统的数据库设计文档和 API 文档。


    状态码约定：为简化设计，除了 `200:OK`（请求成功时）和 `401:Unauthorized`（登录时账号或密码不正确，或请求时 token
    不正确时）外，其他的报错都返回 `422:Unprocessable Entity`，但返回 422 的同时须在 body
    中返回报错信息供客户端显示（若客户端在使用部分接口时，需对都返回 422 的不同结果做区分，则再对相关接口另行设置状态码）。
    
    
    数据库设计：`XXX_ref` 表示一个 Foreign key。
  contact:
    email: yulei2018@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
externalDocs:
  description: View source on GitHub
  url: https://github.com/yusanshi/bank-management-system
servers:
  - url: https://bank.yusanshi.com/api/v1
paths:
  /user/login:
    post:
      tags:
        - user
      summary: 登录，获取 token 和用户信息
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires:
                    type: integer
                    description: 有效期（单位：秒）
        '401':
          description: 账号或密码错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - 账号或密码错误
        '422':
          $ref: '#/components/responses/Unprocessable'
  /user/logout:
    post:
      tags:
        - user
      summary: 退出，使 token 失效。token 无效并不报错
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 退出成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /user/password:
    post:
      tags:
        - user
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: 修改成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /client/:
    get:
      tags:
        - client
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Client'
                    - type: object
                      properties:
                        client_account_associations:
                          type: array
                          items:
                            $ref: '#/components/schemas/ClientAndAccount'
                          
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    post:
      tags:
        - client
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Client'
                - type: object
                  properties:
                    client_account_associations:
                      type: array
                      items:
                        $ref: '#/components/schemas/ClientAndAccount'

      responses:
        '200':
          description: 添加成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /client/{id_number}:
    put:
      tags:
        - client
      parameters:
        - name: id_number
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Client/properties/id_number'
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Client'
                - type: object
                  properties:
                    client_account_associations:
                      type: array
                      items:
                        $ref: '#/components/schemas/ClientAndAccount'

      responses:
        '200':
          description: 修改成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      tags:
        - client
      parameters:
        - name: id_number
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Client/properties/id_number'
        - name: X-Token
          in: header
          schema:
            type: string

      responses:
        '200':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /account/:
    get:
      tags:
        - account
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Account'
                    - type: object
                      properties:
                        client_account_associations:
                          type: array
                          items:
                            $ref: '#/components/schemas/ClientAndAccount'
                            
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    post:
      tags:
        - account
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Account'
                - type: object
                  properties:
                    client_account_associations:
                      type: array
                      items:
                        $ref: '#/components/schemas/ClientAndAccount'

      responses:
        '200':
          description: 添加成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /account/{id}:
    put:
      tags:
        - account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Account/properties/id'
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Account'
                - type: object
                  properties:
                    client_account_associations:
                      type: array
                      items:
                        $ref: '#/components/schemas/ClientAndAccount'

      responses:
        '200':
          description: 修改成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      tags:
        - account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Account/properties/id'
        - name: X-Token
          in: header
          schema:
            type: string

      responses:
        '200':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /loan/:
    get:
      tags:
        - loan
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Loan'
                    - type: object
                      properties:
                        payments:
                          type: array
                          items:
                            $ref: '#/components/schemas/Payment'
                        clients:
                          type: array
                          items:
                            $ref: '#/components/schemas/Client/properties/id_number'
                            
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    post:
      tags:
        - loan
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Loan'
                - type: object
                  properties:
                    payments:
                      type: array
                      items:
                        $ref: '#/components/schemas/Payment'
                    clients:
                      type: array
                      items:
                        $ref: '#/components/schemas/Client/properties/id_number'

      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /loan/{id}:
    put:
      tags:
        - loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Loan/properties/id'
        - name: X-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Loan'

      responses:
        '200':
          description: 修改成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

    delete:
      tags:
        - loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Loan/properties/id'
        - name: X-Token
          in: header
          schema:
            type: string

      responses:
        '200':
          description: 删除成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'

  /bank/:
    get:
      tags:
        - bank
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bank'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'


  /staff/:
    get:
      tags:
        - staff
      parameters:
        - name: X-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Staff'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
          
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Token:
      type: object
      properties:
        token:
          type: string
        expires:
          type: integer
        user_ref:
          $ref: '#/components/schemas/User/properties/username'

    Client:
      type: object
      properties:
        id_number:
          type: string
        name:
          type: string
        phone_number:
          type: string
        address:
          type: string
        contact_name:
          type: string
        contact_phone_number:
          type: string
        contact_email:
          type: string
        contact_relationship:
          type: string
        staff_ref:
          $ref: '#/components/schemas/Staff/properties/id_number'

    Account:
      type: object
      properties:
        id:
          type: string
        balance:
          type: number
        open_date:
          type: integer
        account_type:
          type: string
        bank_ref:
          $ref: '#/components/schemas/Bank/properties/name'


    DepositAccount:
      type: object
      properties:
        id_ref:
          $ref: '#/components/schemas/Account/properties/id'
        interest_rate:
          type: number
        currency_kind:
          type: string

    ChequeAccount:
      type: object
      properties:
        id_ref:
          $ref: '#/components/schemas/Account/properties/id'
        overdraft:
          type: number

    ClientAndAccount:
      type: object
      properties:
        client_ref:
          $ref: '#/components/schemas/Client/properties/id_number'
        account_ref:
          $ref: '#/components/schemas/Account/properties/id'
        last_visit_date:
          type: number

    Staff:
      type: object
      properties:
        id_number:
          type: string
        name:
          type: string
        phone_number:
          type: string
        address:
          type: string
        joined_date:
          type: integer
        kind:
          type: string
        department_ref:
          $ref: '#/components/schemas/Department/properties/id'

    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        kind:
          type: string

    Bank:
      type: object
      properties:
        name:
          type: string
        city:
          type: string

    Loan:
      type: object
      properties:
        id:
          type: integer
        money:
          type: number
        status:
          type: string
        bank_ref:
          $ref: '#/components/schemas/Bank/properties/name'


    Payment:
      type: object
      properties:
        id:
          type: integer
        pay_date:
          type: integer
        money:
          type: number
        loan_ref:
          $ref: '#/components/schemas/Loan/properties/id'

  responses:
    Unauthorized:
      description: token 信息无效
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - token 信息无效

    Unprocessable:
      description: 请求失败（General，应在 message 中附上报错信息）
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string